Add staging buffer before actual vertex buffer.
This is done so that memory of type VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT
can be used and accessed by the GPU. This memory is not accessible from
the CPU. That's why a staging buffer, that the CPU can access, is
created. The data will be moved from staging buffer to the actual vertex
buffer.